rename: StaggeredImageItem 컴포넌트 placelist 패키지로 이동
